requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai==1.61.0
python-dotenv==1.0.0
streamlit==1.41.1
pypdf==5.2.0
faiss-cpu==1.10.0
tiktoken==0.8.0
python-multipart==0.0.9
chromadb==0.4.22
requests==2.32.3
httpx==0.28.1